requirements.txt

numpy>=1.21.0
pandas>=1.3.0
requests>=2.25.0
websocket-client>=1.0.0
scipy>=1.7.0
matplotlib>=3.4.0
python-dateutil>=2.8.0
pytz>=2021.1
PyQt6
pandas
scikit-learn
finnhub-python
numpy
BinaryOptionsToolsV2>=0.1.7